sd-dhcp-server: split into small pieaces
No functional change, just several functions are moved/split/renamed.
