+++ src/libsystemd-network/dhcp-server-lease-internal.h
@@ -5,6 +5,7 @@
 
 #include "dhcp-client-id-internal.h"
 #include "dhcp-server-internal.h"
+#include "dhcp-server-request.h"
 #include "sd-forward.h"
 
 typedef struct sd_dhcp_server_lease {
