+++ src/libsystemd-network/dhcp-server-request.h
@@ -1,0 +1,31 @@
+/* SPDX-License-Identifier: LGPL-2.1-or-later */
+#pragma once
+
+#include "dhcp-client-id-internal.h"
+#include "dhcp-protocol.h"
+#include "sd-forward.h"
+#include "sparse-endian.h"
+#include "time-util.h"
+
+typedef struct DHCPRequest {
+        /* received message */
+        DHCPMessage *message;
+        triple_timestamp timestamp;
+
+        /* options */
+        sd_dhcp_client_id client_id;
+        size_t max_optlen;
+        be32_t server_id;
+        be32_t requested_ip;
+        usec_t lifetime;
+        const uint8_t *agent_info_option;
+        char *hostname;
+        const uint8_t *parameter_request_list;
+        size_t parameter_request_list_len;
+        bool rapid_commit;
+} DHCPRequest;
+
+int dhcp_request_get_lifetime_timestamp(DHCPRequest *req, clockid_t clock, usec_t *ret);
+
+int dhcp_server_handle_message(sd_dhcp_server *server, DHCPMessage *message, size_t length, const triple_timestamp *timestamp);
+int dhcp_server_setup_io_event_source(sd_dhcp_server *server);
