+++ src/libsystemd-network/meson.build
@@ -13,6 +13,8 @@ libsystemd_network_sources = files(
         'dhcp-relay-interface.c',
         'dhcp-relay-upstream.c',
         'dhcp-route.c',
+        'dhcp-server-request.c',
+        'dhcp-server-send.c',
         'dhcp6-network.c',
         'dhcp6-option.c',
         'dhcp6-protocol.c',
